Adds modal component
