+++ vanillajs/modal/index.html
@@ -14,7 +14,7 @@
     <div class="modal-container hide"></div>
   </div>
   <button class="modal-btn">Click me!</button>
-  <script src="app.js"></script>
+  <script src="app.js" type="module"></script>
 </body>
 
 </html>
